requirements.txt
streamlit
Pillow
google-generativeai
openai
pypdf
reportlab
